#4 added some missing terms
This word list is not comprehensive as it does not have all words that differ between the two variants. For instance the words queue (UK) and line (US) are missing. Adding them would cause some problems with the document counts, because "line" has so many other meanings that "queue" does not cover. I added 10 or so word pairs here that shouldn't cause such problems.